Rimossi metodi con tipo di ritorno dinamico dalla classe 'Maths', utilizzato il 'double' come tipo predefinito per il calcolo matematico ed adattate tutte le classi che richiedevano il cast in 'float'.

+++ engine/src/main/java/net/byloth/engine/graphics/TimedShader.java
@@ -82,10 +82,10 @@ public class TimedShader
         }
         else
         {
-            float totalTimeSpan = secondTimedColor.getTime() - firstTimedColor.getTime();
-            float currentTimeSpan = currentTime - firstTimedColor.getTime();
+            int totalTimeSpan = secondTimedColor.getTime() - firstTimedColor.getTime();
+            int currentTimeSpan = currentTime - firstTimedColor.getTime();
 
-            float shadingRatio;
+            double shadingRatio;
 
             if (endToBeginning == true)
             {
